refactor: Address comments pt2
